# Scoreboard-backend/config/swagger.json
{
	"components": {
...
		"schemas": {
...
			"FlattenMaps_T_": {
				"properties": {},
				"type": "object"
			},
			"Document_any.any.any_": {
				"$ref": "#/components/schemas/FlattenMaps_T_",
				"description": "Generic types for Document:\n*  T - the type of _id\n*  TQueryHelpers - Object with any helpers that should be mixed into the Query type\n*  DocType - the type of the actual Document created"
			},
...
			"Required_any_": {
				"properties": {},
				"type": "object",
				"description": "Make all properties in T required"
			},
...
			"Team": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"medals": {
						"properties": {
							"bronze": {
								"type": "number",
								"format": "double"
							},
							"silver": {
								"type": "number",
								"format": "double"
							},
							"gold": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"bronze",
							"silver",
							"gold"
						],
						"type": "object"
					},
					"points": {
						"type": "number",
						"format": "double"
					},
					"logoUrl": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"medals",
					"points"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTeamRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"logoUrl": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTeamRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"logoUrl": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "inter-iit-scoreboard-server",
		"version": "1.0.0",
		"description": "This is the server for the scoreboard for the Inter-IIT-sports-meet made by Karan Gandhi",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Karan-Gandhi"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/admin/teams": {
			"get": {
				"operationId": "GetAllTeams",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Team"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all teams.",
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "AddTeam",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Adds a team.",
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTeamRequest"
							}
						}
					}
				}
			}
		},
		"/api/admin/teams/{name}": {
			"get": {
				"operationId": "GetTeamByName",
				"responses": {
					"200": {
						"description": "The team with the specified name.",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document_any.any.any_"
										},
										{
											"$ref": "#/components/schemas/Team"
										},
										{
											"$ref": "#/components/schemas/Required_any_"
										}
									]
								}
							}
						}
					}
				},
				"description": "Retrieves a team by its name.",
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": [
					{
						"description": "The name of the team.",
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/admin/teams/{id}": {
			"delete": {
				"operationId": "DeleteTeam",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Deletes a team.",
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the team to be deleted.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateTeam",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates a team.",
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the team to update.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTeamRequest"
							}
						}
					}
				}
			}
		},
		"/api/admin/teams/{name}/{medal}": {
			"patch": {
				"operationId": "AddMedal",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Adds a medal to a team.",
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The type of medal to add.",
						"in": "path",
						"name": "medal",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "- The name of the team.",
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}